+++ changelog/mediation/mobileads-bigoads/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 4.2.0.0
+
+### Added
+
+* Added support for BigoADS sdk version 4.2.0
+* Updated minimum supported BigoADS sdk version to 4.2.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.4.0
+
 ## Version 4.0.4.2
 
 ### Added