+++ changelog/mediation/mobileads-chartboost/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 9.6.0.0
+
+### Added
+
+* Added support for Chartboost sdk version 9.6.0
+* Updated minimum supported Chartboost sdk version 9.6.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.4.0
+
 ## Version 9.5.1.1
 
 ### Added