+++ changelog/mediation/mobileads-ironsource/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 7.7.0.0
+
+### Updated
+
+* Added support for IronSource SDK version 7.7.0.0
+* Updated minimum supported IronSource SDK version to 7.7.0.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.4.0
+
 ## Version 7.6.0.0
 
 ### Updated